+++ pkgs/google_cloud_pubsub/example/doc_examples.dart
@@ -26,6 +26,7 @@ Future<void> streamingPullExample(PubSub pubsub) async {
     streamAckDeadlineSeconds: 30,
     retry: RetrySettings(
       maxRetries: 10,
+      totalTimeout: null,
       initialDelay: const Duration(seconds: 1),
       maxDelay: const Duration(seconds: 30),
     ),